Add guardrail metric support.
