+++ packages/back-end/src/controllers/experiments.ts
@@ -299,6 +299,7 @@ export async function postExperiments(
     hypothesis: data.hypothesis || "",
     conversionWindowDays: data.conversionWindowDays || 3,
     metrics: data.metrics || [],
+    guardrails: data.guardrails || [],
     activationMetric: data.activationMetric || "",
     variations: data.variations || [],
     implementation: data.implementation || "code",
@@ -448,6 +449,7 @@ export async function postExperiment(
     "conversionWindowDays",
     "activationMetric",
     "metrics",
+    "guardrails",
     "variations",
     "status",
     "results",
